requirements2.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
asttokens
async-timeout==4.0.3
attrs==23.1.0
backcall
backoff==2.2.1
backports.functools-lru-cache
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.3.29
click==8.1.7
colorama
coloredlogs==15.0.1
comm
dataclasses-json==0.6.1
debugpy
decorator
entrypoints==0.4
exceptiongroup
executing
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata
importlib-resources==6.1.0
ipykernel
ipython
jedi
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client
jupyter_core
langchain==0.0.314
langsmith==0.0.43
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio
ngrok==0.12.0
numpy==1.26.1
onnxruntime==1.16.1
openai==0.28.1
overrides==7.4.0
packaging
pandas==1.5.3
parso
pickleshare
Pillow==10.1.0
platformdirs
posthog==3.0.2
prompt-toolkit
protobuf==3.20.3
psutil
pulsar-client==3.3.0
pure-eval
pyarrow==13.0.0
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments
Pympler==1.0.1
pyngrok==4.1.1
pypdf==3.16.4
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
semver==3.0.2
six
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
stack-data
starlette==0.20.4
streamlit==1.20.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado
tqdm==4.66.1
traitlets
typer==0.9.0
typing-inspect==0.9.0
typing_extensions
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth
websockets==11.0.3
yarl==1.9.2
zipp
